requirements.txt
# Guard Management System Dependencies - Updated for Email-OTP System

# Build tools (avoid wheel build errors)
setuptools<75
wheel

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
motor==3.3.1
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.2.0

# HTTP Client
httpx==0.25.2

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# QR Code Generation
qrcode[pil]==7.4.2
Pillow>=11.0.0,<12


# Date/Time handling
python-dateutil==2.8.2

# Geography and Distance Calculations
geopy==2.4.1

# File handling
aiofiles==23.2.1

# Email sending for OTP
aiosmtplib==3.0.1

# JWT for access and refresh tokens
PyJWT==2.8.0

# Google Drive integration for Excel files
google-api-python-client==2.108.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.1.1

# Background task scheduling
APScheduler==3.10.4

# Rate limiting
slowapi==0.1.9

# Excel/CSV export
openpyxl==3.1.2
pandas>=2.2.3,<2.3

# Timezone handling for IST
pytz==2023.3